refactor(frontend): extract hierarchy row action buttons into their own module
Move the row action-button renderers (renderInlineActions and its
add/planting/delete/more helpers) out of HierarchyColumns into
hierarchyRowActions, and move the shared NameCellCallbacks /
HierarchyColumnOptions types and NON_BLOCKING_TOOLTIP_PROPS constant into a
hierarchyColumnShared module used by both. HierarchyColumns drops the icon
and ContextMenuIndicator imports only the actions used; it shrinks from ~685
to ~403 lines. No behavior change.
